+++ back/src/main/kotlin/com/diploma/data/task/datasource/local/TaskDatabase.kt
@@ -4,6 +4,6 @@ import com.diploma.data.task.TaskEntity
 import com.diploma.data.task.datasource.TaskDataSource
 
 interface TaskDatabase : TaskDataSource {
-    suspend fun saveTask(taskEntity: TaskEntity): Int
-    suspend fun removeTask(id: Int): Boolean
+    suspend fun saveTask(taskEntity: TaskEntity): Long
+    suspend fun removeTask(id: Long): Boolean
 }
